pyproject.toml
[project]
name = "stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.135.1",
    "gunicorn>=25.1.0",
    "logzero>=1.7.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "requests>=2.32.5",
    "smartapi-python>=1.5.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "websocket>=0.2.1",
    "yfinance>=1.2.0",
]
